+++ web/source/settings/style.css
@@ -601,29 +601,31 @@ span.form-info {
 @media screen and (max-width: 60rem) {
 	/* vertical layout */
 	#root {
-		padding: 1rem;
+		padding: 0.5rem;
+		margin: 0;
 		grid-template-columns: 100%;
 		grid-template-rows: auto auto;
 
-		.sidebar {
+		div.sidebar {
 			justify-self: auto;
-			margin-bottom: 2rem;
+			margin-bottom: 0;
 		}
 
-		.sidebar, section.with-sidebar {
+		div.sidebar, section.with-sidebar {
 			border-top-left-radius: $br;
 			border-top-right-radius: $br;
 			border-bottom-left-radius: $br;
 			border-bottom-right-radius: $br;
 		}
 
-		.sidebar a:first-child h2 {
+		section.with-sidebar {
+			grid-column: 1;
+			padding: 1rem;
+		}
+
+		div.sidebar a:first-child h2 {
 			border-top-right-radius: $br;
 		}
-	}
-
-	section {
-		grid-column: 1;
 	}
 
 	.user-profile .overview {
